LayoutTests/fast/dom/SelectorAPI/caseTagX.xhtml

<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE html 
     PUBLIC "-//W3C//DTD XHTML 1.1 Strict//EN"
     "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
    <div style="display: none">
    <span id="lower1">lower</span><br />
    <p id="UPPER1">UPPER</p><br />
    </div>
    <script>
    <![CDATA[ 
        shouldBeNull("document.querySelector('div SPAN')");

        shouldBeFalse("document.getElementById('lower1').webkitMatchesSelector('div SPAN')");
    ]]> 
    </script>
    <script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
