Safari-609.1.11.1/LayoutTests/fast/dom/SelectorAPI/caseTagX-expected.txt

PASS document.querySelector('div SPAN') is null
PASS document.getElementById('lower1').matches('div SPAN') is false
PASS document.getElementById('lower1').webkitMatchesSelector('div SPAN') is false
PASS successfullyParsed is true

TEST COMPLETE
